перенесли render-point в компоненты
