+++ .gitignore
@@ -1,4 +1,4 @@
-# Файлы и папки операционной системы
+﻿# Файлы и папки операционной системы
 .DS_Store
 Thumbs.db
 
@@ -11,4 +11,5 @@ Thumbs.db
 *.log*
 node_modules/
 
-# Файлы сборки (допишите самостоятельно)
+# Файлы сборки (допишите самостоятельно)public/bundle.js
+public/bundle.js.map
